python-face-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0
python-multipart>=0.0.6
pydantic>=2.0.0
